# apis/openapi.yaml
openapi: 3.1.0
info:
  title: Network Inventory Planning API
  version: 1.0.0
  contact:
    name: warehouse-systems
    url: https://github.com/IQVO/network-inventory-planning
  description: |
    Advisory network inventory balancing for the warehouse-systems fleet.

    Network Inventory Planning projects sibling-context facts (site capability,
    site-SKU demand, published capacity) into local read models and proposes
    inter-warehouse transfers. It never mutates physical stock: approval,
    reservation and physical execution live in their owning contexts.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/transfer-proposals:generate:
    post:
      operationId: generateTransferProposals
      tags: [planning]
      summary: Generate advisory transfer proposals from an explicit snapshot.
      description: |
        Deterministic, explainable proposal generation over a caller-provided
        coherent planning snapshot. Validation failures return RFC 7807
        problem+json. Diagnostic boundary: production planning consumes local
        read models via the simulation endpoint instead.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
      responses:
        '200':
          description: Advisory proposals, score-ordered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
  /v1/transfer-simulations:
    get:
      operationId: simulateTransferOptions
      tags: [planning]
      summary: Simulate advisory transfer options from local read models.
      description: |
        Serves an advisory-only simulation built from the service's projected
        network facts. Fail-closed: returns 503 problem+json when the read
        models are unconfigured, incomplete or stale, rather than proposing
        from partial state.
      responses:
        '200':
          description: Advisory simulation from current local read models.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationResponse'
        '503':
          $ref: '#/components/responses/Problem'
  /v1/transfers:approve:
    post:
      operationId: approveTransfer
      tags: [transfers]
      summary: Approve a transfer proposal and start the allocation saga.
      description: |
        Takes a proposal snapshot (origin, destination, SKU, quantity, policy
        version) and validates it against the CURRENT fail-closed read
        models. On success the InterWarehouseTransfer saga persists through
        PROPOSED → APPROVED → ALLOCATING and, in the same transaction,
        enqueue TransferPlanApproved and TransferAllocationRequested into the
        transactional outbox; a relay publishes them to Kafka and
        inventory-storage's reply later drives the transfer to ALLOCATED or
        UNFULFILLABLE. Idempotent per the required Idempotency-Key header:
        replaying the same key with the same payload returns the ORIGINAL
        transfer (replayed: true); the same key with a different payload is a
        409. Fail-closed validation failures (missing/stale/disabled facts)
        return 422; unusable read models return 503.
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: |
            Transfer-level idempotency key; a fresh key per logical approval.
            Replaying a key replays the FIRST transfer it created.
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveTransferRequest'
      responses:
        '200':
          description: |
            The persisted saga state (200, not 201: a replay of the same
            idempotency key returns the original transfer).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApproveTransferResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /v1/rebalance-runs:
    get:
      operationId: listRebalanceRuns
      tags: [rebalance]
      summary: List scheduled rebalance runs, newest first.
      description: |
        The observe-only scheduled-rebalance run history (ADR 0007): one
        row per NIP_REBALANCE_SCHEDULE tick, recording the snapshot
        watermark the pass planned from, the proposal/rejected counts,
        and the outcome — a fail-closed snapshot refusal is a FAILED run
        carrying its reason, never a hidden error. Runs never approve
        anything and never emit an allocation command.

        A COMPLETED run means the planner EVALUATED the network, so a
        COMPLETED run with `proposalCount` 0 genuinely found nothing to
        propose. While no placement-policy / lane catalogue is configured
        (the v1 state) the planner cannot evaluate any transfer, so every
        run is FAILED with a `failClosedReason` naming the missing
        catalogue, and no RebalanceRunCompleted occurrence is published.
        Read FAILED as "nothing could be evaluated", never as "balanced"
        (ADR 0007 Amendment).
      parameters:
        - name: limit
          in: query
          required: false
          description: Page size, clamped to [1, 200]; default 50.
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: The run history, newest first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RebalanceRunsResponse'
        '400':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /v1/transfers:
    get:
      operationId: listTransfers
      tags: [transfers]
      summary: List transfers, newest first.
      description: |
        Read-only page of inter-warehouse transfers from the saga store,
        newest created first, optionally filtered by lifecycle state and by
        origin and/or destination site. Never mutates anything. 400 for an
        unknown state or out-of-range paging; 503 problem+json when the
        instance runs without a database.
      parameters:
        - name: state
          in: query
          required: false
          description: Only transfers currently in this lifecycle state (case-insensitive).
          schema:
            $ref: '#/components/schemas/TransferState'
        - name: originSiteId
          in: query
          required: false
          description: Only transfers leaving this origin site.
          schema:
            $ref: '#/components/schemas/SiteID'
        - name: destinationSiteId
          in: query
          required: false
          description: Only transfers going to this destination site.
          schema:
            $ref: '#/components/schemas/SiteID'
        - name: limit
          in: query
          required: false
          description: Page size, 1 to 200 (default 50).
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          required: false
          description: Number of matching transfers to skip (default 0).
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: One page of transfers.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferList'
        '400':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /v1/transfers/{id}:
    get:
      operationId: getTransfer
      tags: [transfers]
      summary: Get one transfer with its audit trail.
      description: |
        Read-only view of one transfer: its current state and facts plus the
        immutable audit trail of state transitions (timestamp and cause of
        each). 404 problem+json for an unknown id; 503 problem+json when the
        instance runs without a database.
      parameters:
        - name: id
          in: path
          required: true
          description: The transfer id (the transferId returned by approval).
          schema:
            type: string
            minLength: 1
      responses:
        '200':
          description: The transfer and its audit trail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferDetail'
        '404':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
  /v1/transfers/{id}:cancel:
    post:
      operationId: cancelTransfer
      tags: [transfers]
      summary: Cancel a transfer that has not yet been released (operator action).
      description: |
        The operator's PRE-RELEASE cancel (ADR 0011, ADR 0003). Legal only
        while the transfer is DRAFT, PROPOSED, APPROVED or ALLOCATING; once
        an origin reservation exists (ALLOCATED) or the transfer is past it,
        the request is a 409 `transfer-not-cancellable` and nothing changes.
        The new CANCELLED state, its audit entry (cause = `reason`) and the
        TransferStateAdvanced analytics occurrence commit in one
        transaction. No integration event is emitted. Idempotent: cancelling
        an already CANCELLED transfer is a 200 no-op that returns the
        transfer unchanged. A blank `reason` is a 422 `invalid-cancellation`;
        an unknown id is a 404 `transfer-not-found`; 503 problem+json when the
        instance runs without persistence and the outbox. REST only: the MCP
        surface stays read-only.
      parameters:
        - name: id
          in: path
          required: true
          description: The transfer id (the transferId returned by approval).
          schema:
            type: string
            minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelTransferRequest'
      responses:
        '200':
          description: The transfer, now CANCELLED (also returned for an already CANCELLED transfer).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferView'
        '400':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
        '422':
          $ref: '#/components/responses/Problem'
        '503':
          $ref: '#/components/responses/Problem'
...
components:
...
  responses:
    Problem:
      description: RFC 7807 problem+json error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          description: Stable machine-readable problem identifier.
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
    SiteID:
      type: string
      minLength: 1
      description: Facility-layout Site code identifying one warehouse.
    SKU:
      type: string
      minLength: 1
    Position:
      type: object
      required: [Site, SKU, Available, AsOf]
      properties:
        Site:
          $ref: '#/components/schemas/SiteID'
        SKU:
          $ref: '#/components/schemas/SKU'
        Available:
          type: integer
          minimum: 0
        CustomerReservations:
          type: integer
          minimum: 0
        ConfirmedInbound:
          type: integer
          minimum: 0
        CommittedOutbound:
          type: integer
          minimum: 0
        AsOf:
          type: string
          format: date-time
    Policy:
      type: object
      required: [Version, Site, SKU]
      properties:
        Version:
          type: string
        Site:
          $ref: '#/components/schemas/SiteID'
        SKU:
          $ref: '#/components/schemas/SKU'
        SafetyStock:
          type: integer
          minimum: 0
        TargetStock:
          type: integer
          minimum: 0
        UnitPriority:
          type: integer
    Lane:
      type: object
      required: [Origin, Destination, Enabled]
      properties:
        Origin:
          $ref: '#/components/schemas/SiteID'
        Destination:
          $ref: '#/components/schemas/SiteID'
        LeadTime:
          type: integer
          description: Lead time in nanoseconds (Go time.Duration wire shape).
        UnitHandlingCost:
          type: integer
        Enabled:
          type: boolean
    GenerateRequest:
      type: object
      required: [asOf, positions, policies, lanes]
      properties:
        asOf:
          type: string
          format: date-time
        positions:
          type: array
          items:
            $ref: '#/components/schemas/Position'
        policies:
          type: array
          items:
            $ref: '#/components/schemas/Policy'
        lanes:
          type: array
          items:
            $ref: '#/components/schemas/Lane'
    Proposal:
      type: object
      required: [Origin, Destination, SKU, Quantity, PolicyVersion, PositionAsOf, Reasons, ScoreBreakdown]
      properties:
        Origin:
          $ref: '#/components/schemas/SiteID'
        Destination:
          $ref: '#/components/schemas/SiteID'
        SKU:
          $ref: '#/components/schemas/SKU'
        Quantity:
          type: integer
          minimum: 1
        PolicyVersion:
          type: string
        PositionAsOf:
          type: string
          format: date-time
        Reasons:
          type: array
          items:
            type: string
            enum: [DESTINATION_BELOW_TARGET, ORIGIN_ABOVE_SAFETY_STOCK, APPROVED_LANE]
        ScoreBreakdown:
          type: object
          required: [PriorityBenefit, HandlingPenalty, LeadTimePenalty]
          properties:
            PriorityBenefit:
              type: integer
            HandlingPenalty:
              type: integer
            LeadTimePenalty:
              type: integer
    GenerateResponse:
      type: object
      required: [proposals]
      properties:
        proposals:
          type: array
          items:
            $ref: '#/components/schemas/Proposal'
    SimulationResponse:
      type: object
      required: [advisory, asOf, sites]
      properties:
        advisory:
          type: boolean
          description: Always true. The simulation reserves and moves nothing.
        asOf:
          type: string
          format: date-time
          description: Oldest watermark among the facts the simulation used.
        sites:
          type: array
          description: |
            One advisory view per site, built from the local read models. This
            endpoint does NOT propose transfers; proposals come from
            POST /v1/transfer-proposals:generate over an explicit snapshot, and
            an operator approves one with POST /v1/transfers:approve.
          items:
            $ref: '#/components/schemas/SiteSimulation'
    SiteSimulation:
      type: object
      required:
        - site
        - originEnabled
        - destinationEnabled
        - totalDemand
        - capacityOverWindow
        - capacityHeadroom
        - windowStart
        - windowEnd
      properties:
        site:
          type: string
        originEnabled:
          type: boolean
          description: The site may originate transfers.
        destinationEnabled:
          type: boolean
          description: The site may receive transfers.
        totalDemand:
          type: integer
          description: Demand over the plan window.
        capacityOverWindow:
          type: number
          description: Published capacity over the plan window.
        capacityHeadroom:
          type: integer
          description: Capacity minus demand. Negative means the site is short.
        windowStart:
          type: string
          format: date-time
        windowEnd:
          type: string
          format: date-time
    ApproveTransferRequest:
      type: object
      required: [originSiteId, destinationSiteId, sku, quantity, policyVersion, proposalAsOf]
      properties:
        originSiteId:
          $ref: '#/components/schemas/SiteID'
        destinationSiteId:
          $ref: '#/components/schemas/SiteID'
        sku:
          $ref: '#/components/schemas/SKU'
        quantity:
          type: integer
          minimum: 1
        policyVersion:
          type: string
          minLength: 1
          description: The policy version the advisory proposal was scored under.
        operatorReason:
          type: string
          description: The operator's captured approval reason (audit trail).
        proposalAsOf:
          type: string
          format: date-time
          description: The advisory proposal snapshot's as-of watermark.
    ApproveTransferResponse:
      type: object
      required: [transferId, state, originSiteId, destinationSiteId, sku, quantity, policyVersion, replayed, expiresAt, transferLineId]
      properties:
        transferId:
          type: string
          description: The saga aggregate id.
        state:
          type: string
          enum: [DRAFT, PROPOSED, APPROVED, ALLOCATING, ALLOCATED, PICKED, IN_TRANSIT, ARRIVED, RECEIVED, UNFULFILLABLE, CANCELLED]
          description: |
            The persisted saga state at response time — ALLOCATING
            immediately after a fresh approval (awaiting inventory-storage's
            reply). PICKED/IN_TRANSIT/ARRIVED/RECEIVED are driven later by
            fulfillment-execution's transfer facts and inventory-storage's
            destination facts (ADR 0005); RECEIVED is terminal.
        originSiteId:
          $ref: '#/components/schemas/SiteID'
        destinationSiteId:
          $ref: '#/components/schemas/SiteID'
        sku:
          $ref: '#/components/schemas/SKU'
        quantity:
          type: integer
        policyVersion:
          type: string
        reservationId:
          type: string
          description: inventory-storage's reservation id, present once ALLOCATED.
        rejectionReason:
          type: string
          enum: [ORIGIN_SITE_UNKNOWN, INSUFFICIENT_USABLE, IDEMPOTENCY_CONFLICT]
          description: The closed rejection reason, present once UNFULFILLABLE.
        replayed:
          type: boolean
          description: True when the idempotency key already existed (the original transfer is returned).
        expiresAt:
          type: string
          format: date-time
          description: The approval's expiry deadline (24h in v1).
        transferLineId:
          type: string
          description: |
            The v1 single transfer line id, "<transferId>:1" — the
            inventory-storage allocation ledger key and the reply subject.
...
    RebalanceRunsResponse:
      type: object
      required: [runs]
      properties:
        runs:
          type: array
          items:
            $ref: '#/components/schemas/RebalanceRun'
    TransferState:
      type: string
      enum: [DRAFT, PROPOSED, APPROVED, ALLOCATING, ALLOCATED, PICKED, IN_TRANSIT, ARRIVED, RECEIVED, UNFULFILLABLE, CANCELLED]
      description: |
        Saga lifecycle state. RECEIVED, UNFULFILLABLE and CANCELLED are
        terminal; every other state is waiting for a next step.
    TransferView:
      type: object
      required: [id, state, sku, quantity, originSiteId, destinationSiteId, policyVersion, expiresAt, createdAt, updatedAt, version]
      properties:
        id:
          type: string
          description: The transfer (saga aggregate) id.
        state:
          $ref: '#/components/schemas/TransferState'
        sku:
          $ref: '#/components/schemas/SKU'
        quantity:
          type: integer
          description: Units planned to move.
        pickedQuantity:
          type: integer
          description: Units the origin pick actually picked; present from PICKED onwards (may be short of quantity).
        originSiteId:
          $ref: '#/components/schemas/SiteID'
        destinationSiteId:
          $ref: '#/components/schemas/SiteID'
        policyVersion:
          type: string
        reservationId:
          type: string
          description: inventory-storage's reservation id, present once ALLOCATED.
        rejectionReason:
          type: string
          enum: [ORIGIN_SITE_UNKNOWN, INSUFFICIENT_USABLE, IDEMPOTENCY_CONFLICT]
          description: The closed rejection reason, present once UNFULFILLABLE.
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          description: Time of the last state transition; a non-terminal transfer whose updatedAt is old is stuck.
        version:
          type: integer
          format: int64
          description: Optimistic-concurrency version (the number of persisted state transitions).
...
    TransferDetail:
      allOf:
        - $ref: '#/components/schemas/TransferView'
        - type: object
          required: [audit]
          properties:
            audit:
              type: array
              description: The immutable state-transition trail, oldest first.
              items:
                $ref: '#/components/schemas/AuditEntry'
    CancelTransferRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason:
          type: string
          minLength: 1
          pattern: '\S'
          description: The operator's justification; required and not blank. It becomes the cause of the TransferCancelled audit entry.
    TransferList:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransferView'
        total:
          type: integer
          description: Number of transfers matching the filter before paging.
        limit:
          type: integer
        offset:
          type: integer
...